split merman
